1.3.8.2 SG

+++ basic/modules/photogallery/controllers/admin/ImageController.php
@@ -97,7 +97,7 @@ class ImageController extends Controller
           } */
 
         $categoriesCount = Yii::$app->db->createCommand("SELECT COUNT(*) FROM category")->queryOne();
-        if($categoriesCount['COUNT(*)'] == 0){
+        if ($categoriesCount['COUNT(*)'] == 0) {
             Yii::$app->session->setFlash("warning", "To upload an image, you must create at least one category!");
         }
         return $this->render('create', [
@@ -135,16 +135,16 @@ class ImageController extends Controller
     public function actionDelete($id)
     {
         $image = $this->findModel($id);
-        if (isset($image->image)) {
-            unlink(Url::to("@app" . Yii::getAlias($image->image)));
+        if ($image->delete()) {
+            if (isset($image->image)) {
+                unlink(Url::to("@app" . Yii::getAlias($image->image)));
+            }
+
+            $imageCategory = \app\modules\photogallery\models\Category::findOne(['slug' => $image->category]);
+            $imagesCount = Yii::$app->db->createCommand("SELECT COUNT(*) FROM image WHERE category = '$image->category'")->queryOne();
+            $imageCategory->count = $imagesCount['COUNT(*)'];
+            $imageCategory->save();
         }
-        $image->delete();
-
-        $imageCategory = \app\modules\photogallery\models\Category::findOne(['slug' => $image->category]);
-        $imagesCount = Yii::$app->db->createCommand("SELECT COUNT(*) FROM image WHERE category = '$image->category'")->queryOne();
-        $imageCategory->count = $imagesCount['COUNT(*)'];
-        $imageCategory->save();
-
 
         return $this->redirect(Url::toRoute(['admin/image/index', 'slug' => $image->category]));
     }
